python/requirements.txt
pylightxl
pyinstaller
ruamel.yaml